--- frontend/app/farcaster/[name]/page.tsx
@@ -1,46 +1,0 @@
-'use client'
-
-import { useEffect, useState } from "react";
-import sdk, { type FrameContext } from "@farcaster/frame-sdk";
-
-export default function Farcaster() {
-  const [isSDKLoaded, setIsSDKLoaded] = useState(false);
-  const [context, setContext] = useState<FrameContext>();
-
-  useEffect(() => {
-    const load = async () => {
-      const context = await sdk.context;
-      setContext(context);
-
-      sdk.actions.ready({});
-    };
-
-    if (sdk && !isSDKLoaded) {
-      setIsSDKLoaded(true);
-      load();
-      return () => {
-        sdk.removeAllListeners();
-      };
-    }
-  }, [isSDKLoaded]);
-
-  if (!isSDKLoaded) {
-    return <div>Loading...</div>;
-  }
-
-  return (
-    <div style={{ 
-      paddingTop: context?.client.safeAreaInsets?.top ?? 0,
-      paddingBottom: context?.client.safeAreaInsets?.bottom ?? 0,
-      paddingLeft: context?.client.safeAreaInsets?.left ?? 0,
-      paddingRight: context?.client.safeAreaInsets?.right ?? 0,
-    }}>
-      <div>
-        Client FID: {context?.client.clientFid}<br/>
-        Display Name: {context?.user.displayName}<br/> 
-        Username: {context?.user.username}<br/>
-        FID: {context?.user.fid}
-      </div>
-    </div>
-  );
-}
